защита правки
